+++ frontend/src/App.jsx
@@ -1,10 +1,6 @@
 import React from 'react';
-import { useState, useEffect } from 'react';
-import {
-  Route,
-  Routes,
-  useNavigate,
-} from "react-router-dom";
+import { useContext, useEffect } from 'react';
+import {Route, Routes,} from "react-router-dom";
 import Layout from "./components/Layout/Layout";
 import Homepage from "./views/Homepage";
 import Login from "./components/Login/Login";
@@ -13,21 +9,16 @@ import MoodTracker from "./components/MoodTracker/MoodTracker";
 import TravelMood from "./components/TravelMood/TravelMood";
 import DestinationDetail from "./components/DestinationView/DestinationView";
 import PageNotFound from "./views/PageNotFound";
-
-
+import { Context } from './context/Context.jsx';
 // App component with all the routing logic for make posible the navigation between the different views and keeping the Navbar component always visible at the top of the page 
 const App = () => {
-  const [isAuthenticated, setIsAuthenticated] = useState(false);
-  const [username, setUsername] = useState("");
-  const [userId, setUserId] = useState(null);
-  const [userImage, setUserImage] = useState(null);
-  const navigate = useNavigate();
+
+  const {  setUsername, setUserId, setUserImage, setIsAuthenticated} = useContext(Context)
 
   useEffect(() => {
     const storedUsername = localStorage.getItem('username');
     const storedUserId = localStorage.getItem('userId');
     const storedUserImage = localStorage.getItem('userImage');
-
     if (storedUsername && storedUserId && storedUserImage) {
       setUsername(storedUsername);
       setUserId(storedUserId);
@@ -36,26 +27,16 @@ const App = () => {
     }
   }, []);
 
-  const handleLogout = () => {
-    localStorage.removeItem('username');
-    localStorage.removeItem('userId');
-    localStorage.removeItem('userImage');
-    setUsername('');
-    setUserId(null);
-    setUserImage(null);
-    setIsAuthenticated(false);
-    navigate('/');
-  };
 
   return (
     <Routes>
-      <Route path="/" element={<Layout isAuthenticated={isAuthenticated} userImage={userImage} username={username} onLogout={handleLogout} />}>
-        <Route index element={<Homepage isAuthenticated={isAuthenticated} />} />
-        <Route path="/mood-tracker" element={<MoodTracker userId={userId} />} />
+      <Route path="/" element={<Layout />}>
+        <Route index element={<Homepage  />} />
+        <Route path="/mood-tracker" element={<MoodTracker />} />
         <Route path="/travel-mood" element={<TravelMood />} />
         <Route path="/destination/:name" element={<DestinationDetail />} />
-        <Route path="/login" element={<Login setUserId={setUserId} setIsAuthenticated={setIsAuthenticated} setUsername={setUsername} />} />
-        <Route path="/register" element={<Register setUserId={setUserId} setIsAuthenticated={setIsAuthenticated} setUsername={setUsername} />} />
+        <Route path="/login" element={<Login  />} />
+        <Route path="/register" element={<Register  />} />
         <Route path="*" element={<PageNotFound />} />
       </Route>
     </Routes>
