+++ frontend/src/components/Layout/Layout.jsx
@@ -2,16 +2,11 @@ import React from 'react'
 import { Outlet } from 'react-router-dom'
 import Navbar from '../Navbar/Navbar'
 
-function Layout({ isAuthenticated, userImage, username, onLogout }) {
+function Layout() {
   return (
     <>
       <header>
         <Navbar 
-          isAuthenticated={isAuthenticated}
-          userImage={userImage}
-          defaultProfileImage={null} // Add the default profile image here
-          username={username}
-          onLogout={onLogout}
         />
       </header>
       <main>
