+++ frontend/src/components/MoodTracker/MoodTracker.jsx
@@ -1,155 +1,17 @@
-import React, { useState, useEffect } from 'react';
+import React, { useContext, useEffect } from 'react';
 import './MoodTracker.css';
+import { Context } from '../../context/Context.jsx';
+const MoodTracker = () => {
 
-const MoodTracker = ({ userId }) => {
-  const [selectedMood, setSelectedMood] = useState(null);
-  const [moodLog, setMoodLog] = useState([]);
-  const [recommendations, setRecommendations] = useState([]);
-  const [edit, setEdit] = useState(false);
-
-
-  // function to get user Mood:
-  const fetchMoodLog = async () => {
-    try {
-      const response = await fetch(`http://localhost:4000/users/${userId}`, {
-        method: 'GET',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        credentials: 'include',
-      });
-      if (!response.ok) {
-        const errorData = await response.json();
-        throw new Error(errorData.error.message);
-      }
-      const data = await response.json();
-      setMoodLog(data.moods);
-      localStorage.setItem('userMood', JSON.stringify(data.moods));
-    } catch (err) {
-      console.log(err.message);
-    }
-  };
+  const {
+    userId, selectedMood, moodLog, recommendations, edit, setEdit,
+    handleLogMood, handleDeleteMood, handleMoodSelect, handleReplaceMood,
+    fetchMoodLog, handleGetRecommendations} = useContext(Context)
 
   // useEffect to fetch users mood every time a new user logs in:
   useEffect(() => {
     fetchMoodLog();
   }, [userId]);
-
-
-
-  // function to select a mood:
-  const handleMoodSelect = (mood) => {
-    if (selectedMood === mood) {
-      setSelectedMood(null);
-    } else {
-      setSelectedMood(mood)
-    }
-
-  };
-
-  // function to log the mood:
-  const handleLogMood = async () => {
-    if (!selectedMood) {
-      alert("Please select a mood before logging")
-    } else {
-
-      try {
-        const response = await fetch(`http://localhost:4000/users/${userId}/moods`, {
-          method: 'PATCH',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          credentials: 'include',
-          body: JSON.stringify({ selectedMood }),
-        });
-
-        if (!response.ok) {
-          const errorData = await response.json();
-          throw new Error(errorData.error.message);
-        }
-        await fetchMoodLog();
-        setSelectedMood(null);
-        setEdit(false);
-      } catch (err) {
-        setSelectedMood(null);
-        alert(err.message);
-      }
-    }
-  };
-
-  // funtion to delete mood:
-  const handleDeleteMood = async (moodId) => {
-    try {
-      const response = await fetch(`http://localhost:4000/users/${userId}/${moodId}`, {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        credentials: 'include',
-      });
-
-      if (!response.ok) {
-        const errorData = await response.json();
-        throw new Error(errorData.error.message);
-      }
-      await fetchMoodLog();
-      setSelectedMood(null)
-    } catch (err) {
-      alert(err.message);
-    }
-  };
-
-  // function to replace mood:
-  const handleReplaceMood = async (moodId) => {
-    try {
-      if (!selectedMood) {
-        alert("Please select a mood to Replace with")
-      } else {
-        const response = await fetch(`http://localhost:4000/users/${userId}/${moodId}`, {
-          method: 'PATCH',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          credentials: 'include',
-          body: JSON.stringify({ selectedMood }),
-        });
-
-        if (!response.ok) {
-          const errorData = await response.json();
-          throw new Error(errorData.error.message);
-        }
-        await fetchMoodLog();
-        setSelectedMood(null)
-        
-      }
-    } catch (err) {
-      alert(err.message);
-    }
-  };
-
-
-  
- 
-  const handleGetRecommendations = async () => {
-    try {
-      const response = await fetch('http://localhost:4000/travel', {
-        method: 'GET',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        credentials: 'include',
-      });
-
-      if (!response.ok) {
-        throw new Error(`Error getting recommendations: ${response.statusText}`);
-      }
-
-      const data = await response.json();
-      setRecommendations(data);
-    } catch (error) {
-      console.error('Error getting recommendations:', error);
-    }
-  };
 
   return (
     <div className="MoodTrackerContainer">
@@ -172,7 +34,6 @@ const MoodTracker = ({ userId }) => {
         <button className="ActionButton" onClick={handleGetRecommendations}>Get Recommendations</button>
       </div>
 
-
       <div className="MoodLogContainer">
         <h3>Mood Log</h3>
         {moodLog.length > 0 && moodLog.map((entry) => {
@@ -191,8 +52,9 @@ const MoodTracker = ({ userId }) => {
         })}
       </div>
 
+      <h3>Recommendations</h3>
 
-      <div className="RecommendationsContainer">
+      {/* <div className="RecommendationsContainer">
         <h3>Recommendations</h3>
         {recommendations.map((rec, index) => (
           <div key={index}>
@@ -211,7 +73,7 @@ const MoodTracker = ({ userId }) => {
             ))}
           </div>
         ))}
-      </div>
+      </div> */}
     </div>
   );
 };
