+++ frontend/src/components/Register/Register.jsx
@@ -1,86 +1,15 @@
-import React, { useState, useRef } from "react";
+import React, { useContext } from "react";
 import ReCAPTCHA from "react-google-recaptcha";
-import { useNavigate } from "react-router-dom";
 import "./Register.css";
-import defaultProfileImage from '../../assets/default-profile.png';
+import { Context } from "../../context/Context.jsx";
 
-const Register = ({ setUserId, setIsAuthenticated }) => {
-  const [email, setEmail] = useState("");
-  const [username, setUsername] = useState("");
-  const [password, setPassword] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
-  const [recaptchaToken, setRecaptchaToken] = useState("");
-  const [profileImage, setProfileImage] = useState("");
-  const [error, setError] = useState("");
-  const [success, setSuccess] = useState("");
-  const navigate = useNavigate();
-  const fileInput = useRef(null); 
-  const recaptchaRef = useRef(null);
+const Register = () => {
 
-  const handleRegister = async (e) => {
-    e.preventDefault();
-    setError("");
-    setSuccess("");
+  const { email, setEmail, username, password, error, setUsername, 
+    setPassword, confirmPassword, setConfirmPassword, 
+    setRecaptchaToken, setProfileImage, fileInput, recaptchaRef, 
+    handleRegister } = useContext(Context)
 
-    if (password !== confirmPassword) {
-      setError("Passwords do not match");
-      return;
-    }
-
-    if (!recaptchaToken) {
-      alert("Please complete the reCAPTCHA");
-      return;
-    }
-
-    try {
-      const formData = new FormData();
-      formData.append('email', email);
-      formData.append('username', username);
-      formData.append('password', password);
-      formData.append('profileImage', profileImage);
-      formData.append('recaptchaToken', recaptchaToken);
-    
-      const settings = {
-        method: "POST",
-        credentials: "include",
-        body: formData,
-      }
-
-      const response = await fetch(`${import.meta.env.VITE_API}/register`, settings);
-
-      if (response.ok) {
-        const data = await response.json();
-
-        
-        localStorage.setItem('username', data.username);
-        localStorage.setItem('userId', data.id);
-        localStorage.setItem('userImage', data.profileImage || defaultProfileImage);
-        setUserId(data.id);
-        setIsAuthenticated(true);
-        setUsername(data.username);
-        setProfileImage(data.profileImage || defaultProfileImage);
-        
-        fileInput.current.value = ""; 
-        recaptchaRef.current.reset();
-        setSuccess("Registration successful");
-        setRecaptchaToken("");
-
-        alert(`Registration successful. Welcome, ${data.username}!`);
-
-
-        navigate('/');
-        refreshPage();
-      } else {
-        const errorData = await response.json();
-        console.log(errorData)
-        setError(errorData.message || 'Registration failed');
-        alert(errorData.error.message);
-      }
-    } catch (error) {
-      setError("An error occurred");
-      alert('An error occurred');
-    }
-  }
 
   const siteKey = import.meta.env.VITE_RECAPTCHA_SITE_KEY;
   console.log("ReCAPTCHA Site Key:", siteKey); // Debugging line
@@ -126,12 +55,12 @@ const Register = ({ setUserId, setIsAuthenticated }) => {
           />
         </label>
         <label>
-            Profile Image:
-            <input
-                type="file"
-                onChange={(e) => setProfileImage(e.target.files[0])}
-                ref={fileInput}
-            />
+          Profile Image:
+          <input
+            type="file"
+            onChange={(e) => setProfileImage(e.target.files[0])}
+            ref={fileInput}
+          />
         </label>
         {siteKey && (
           <ReCAPTCHA
@@ -141,7 +70,7 @@ const Register = ({ setUserId, setIsAuthenticated }) => {
           />
         )}
         {error && <p className="error">{error}</p>}
-        {success && <p className="success">{success}</p>}
+
         <button type="submit">Register</button>
       </form>
     </div>
