+++ frontend/src/main.jsx
@@ -3,11 +3,16 @@ import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
 import { BrowserRouter } from 'react-router-dom'
 import './index.css'
+import ContextProvider from './context/Context.jsx'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <BrowserRouter>
-      <App />
+
+      <ContextProvider>
+        <App />
+      </ContextProvider>
+
     </BrowserRouter>
   </React.StrictMode>,
 )
